requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==47.0.0
Django==4.2.16
django-anymail==10.0
django-cors-headers==4.3.1
djangorestframework==3.16.1
filelock==3.20.3
fsspec==2026.1.0
hf-xet==1.4.2
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pillow==11.0.0
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
sqlparse==0.5.5
sympy==1.14.0
tokenizers==0.19.1
torch==2.1.2
tqdm==4.67.1
transformers==4.41.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
